完成问题的阅读数，评论功能

+++ src/main/java/community/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         Question question=questionMapper.selectByPrimaryKey(id);
         model.addAttribute("id",id);
@@ -40,7 +40,7 @@ public class PublishController {
 
     @PostMapping("/publish")
     public String doPublish(
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
@@ -75,7 +75,6 @@ public class PublishController {
         question.setDescription(description);
         question.setTag(tag);
         question.setCreator(user.getId());
-
         questionService.createOrUpdate(question);
         return "redirect:/";
     }
